revise sale report

+++ src/main/java/beans/Food.java
@@ -2,7 +2,7 @@ package beans;
 
 public class Food extends MenuItem{
 
-    public Food(long id, String name, Type type, String description, double price){
+    public Food(int id, String name, Type type, String description, double price){
         super(id, name, type, description, price);
     }
